VPAAMP-683 : add log to to check time taken between first frame to pipeline moved to PLAYING state
Reason for change :A new log has been added to report the time elapsed between the first frame and the PLAYING state
Test Procedure: See ticket
Risks: low

+++ middleware/InterfacePlayerRDK.h
@@ -163,7 +163,7 @@ class InterfacePlayerRDK
 		void *mDRMSessionManager;
 		std::map<InterfaceCB, std::function<void()>> callbackMap;
 		std::map<InterfaceCB, std::function<void(int)>> setupStreamCallbackMap;
-		long long mFirstFrameTimeInMS;
+		long mFirstFrameTimeInMS;
         
 		PlayerScheduler mScheduler;
 		InterfacePlayerRDK(bool isRialto = false);
